add security group test cases
